WiFi Network Scanner/requirements.txt
tk
matplotlib
qrcode
Pillow
fpdf

